% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{dmatrix.spearman.dissimilarity}
\alias{dmatrix.spearman.dissimilarity}
\title{dmatrix.spearman.dissimilarity}
\usage{
dmatrix.spearman.dissimilarity(X)
}
\arguments{
\item{X}{A numeric matrix with feature in colnames}
}
\value{
Return a dissimilarity matrix of size PxP.
}
\description{
Compute the spearman dissimilarity distance.
}
