% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUpDownCluster.R
\name{.apply_fc_threshold}
\alias{.apply_fc_threshold}
\title{demo <- suppressWarnings(get_demo_cluster())
x <- diff(demo$X[,1])
diff_threshold <- 15}
\usage{
.apply_fc_threshold(x, diff_threshold)
}
\description{
demo <- suppressWarnings(get_demo_cluster())
x <- diff(demo$X[,1])
diff_threshold <- 15
}
