% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCluster.R
\name{get_demo_cluster}
\alias{get_demo_cluster}
\title{get_demo_cluster}
\usage{
get_demo_cluster()
}
\value{
a list containg:
\item{X}{data.frame}
\item{Y}{data.frame}
\item{Z}{data.frame}
\item{pca}{a mixOmics pca result}
\item{spca}{a mixOmics spca result}
\item{pls}{a mixOmics pls result}
\item{spls}{a mixOmics spls result}
\item{block.pls}{a mixOmics block.pls result}
\item{block.spls}{a mixOmics block.spls result}
}
\description{
Generates random data to be used in examples.
}
\examples{
# Random data could lead to "The SGCCA algorithm did not converge" warning which is not important for a demo
demo <- suppressWarnings(get_demo_cluster())
}
