% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{get_demo_silhouette}
\alias{get_demo_silhouette}
\title{Get data for silhouette demo}
\usage{
get_demo_silhouette()
}
\value{
A matrix of expression profile, sample in raws, time in columns.
}
\description{
Get data for silhouette demo
}
\examples{
data <- get_demo_silhouette()

}
