% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNcomp.R
\name{getNcomp}
\alias{getNcomp}
\title{Get optimal number of components}
\usage{
getNcomp(object, max.ncomp = NULL, X, Y = NULL, ...)
}
\arguments{
\item{object}{A mixOmics object of the class `pca`, `spca`, `mixo_pls`, `mixo_spls`, `block.pls`, `block.spls`}

\item{max.ncomp}{integer, minimum number of component to include.
If no argument is given, `max.ncomp=object$ncomp`}

\item{min.ncomp}{integer, minimum number of component to include.
Must be greater than 0. If no argument is given, `min.ncomp = 1`.}
}
\value{
\code{getNcomp} returns a list with class "ncomp.tune.silhouette" containing the following components:

\item{ncomp}{a vector containing the tested ncomp}
\item{silhouette}{a vector containing the average silhouette coefficient by ncomp}
\item{dmatrix}{the distance matrix used to compute silhouette coefficient}
}
\description{
Compute the average silhouette coefficient for a given set of components on a mixOmics result.
Foreach given ncomp, the mixOmics method is performed with the sames arguments and the given `ncomp`.
Longitudinal clustering is performed and average silhouette coefficient is computed.
}
\examples{
# random input data
X <- matrix(rnorm(1:100), ncol = 10, dimnames = list(1:10, paste0("X_", 1:10)))

# Principal Component Analysis
pca.res <- pca(X, ncomp = 5)

# getNcomp
res.ncomp <- getNcomp(pca.res, max.ncomp = 4)
plot(res.ncomp)

}
\seealso{
\code{\link{getCluster}}, \code{\link{silhouette}}
}
