% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmSpline-method.R
\docType{class}
\name{lmms-class}
\alias{lmms-class}
\title{\code{lmms} class a S4 superclass to extend \code{lmmspline}  and \code{lmmsde} class.}
\description{
\code{lmms} class is a superclass for classes \code{lmmspline}  and \code{lmmsde}. These classes inherit common slots.
}
\section{Slots}{

\describe{
\item{\code{basis}}{An object of class \code{character} describing the basis used for modelling.}

\item{\code{knots}}{An object of class \code{numeric}, describing the boundaries of the splines. If not defined or if basis='cubic' knots are automatically estimated using Ruppert 2002 or are the design points when using 'cubic'.}

\item{\code{errorMolecules}}{Vector of class \code{character}, describing the molecules that could not be modelled.}
}}

