% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmSpline-method.R
\docType{class}
\name{lmmspline-class}
\alias{lmmspline-class}
\title{\code{lmmspline} class a S4 class that extends \code{lmms} class.}
\description{
\code{lmmspline} class inherits from class \code{lmms} and extends it with three further slots: \code{predSpline}, \code{modelsUsed}, \code{models}. The class is returned when applying \code{\link{lmmSpline}} method.
}
\section{Slots}{

\describe{
\item{\code{predSpline}}{A \code{data.frame} returning the fitted values for the time points of interest.}

\item{\code{models}}{A \code{list} of class \code{\link{lm}} or  \code{\link{lme}} containing the models for every molecule}

\item{\code{modelsUsed}}{A \code{list} of class \code{lm} or \code{lme}, containing the models used to model the particular feature of interest.}

\item{\code{derivative}}{A \code{logical} value indicating if the derivative was calculated.}
}}

