% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proprotionality.R
\name{proprortionality}
\alias{proprortionality}
\title{Proportionality Distance}
\usage{
proprortionality(X)
}
\arguments{
\item{X}{an object of the class: \code{pca}, \code{spca}, \code{pls}, \code{spls}, \code{block.pls}, \code{block.spls}}
}
\value{
Return a list containing the following components:
\describe{
  \item{propr.distance}{Square matrix with proprortionality distance between pairs of features}
  \item{pvale}{Wilcoxon U-test p-value comparing the medians within clusters and with the entire background set}
}
}
\description{
\code{proprortionality} is a wrapper that compute proportionality distance for 
a clustering result (\code{pca}, \code{spca}, \code{pls}, \code{spls}, \code{block.pls}, \code{block.spls}).
and it performs a u-test to compare the median within a cluster to the median of the entire background set.
}
\examples{
demo <- suppressMessages(get_demo_cluster())
X <- demo$pca
proprortionality(X)
X <- demo$spls
proprortionality(X)
X <- demo$block.spls
proprortionality(X)

}
\references{
Lovell, D., Pawlowsky-Glahn, V., Egozcue, J. J., Marguerat, S., Bähler, J. (2015). Proportionality: a valid alternative to correlation for relative data. PLoS Comput. Biol. 11, e1004075. doi: 10.1371/journal.pcbi.1004075
Quinn, T. P., Richardson, M. F., Lovell, D., Crowley, T. M. (2017). propr: an r-package for identifying proportionally abundant features using compositional data analysis. Sci. Rep. 7, 16252. doi: 10.1038/s41598-017-16520-0
}
\seealso{
\code{\link[propr]{propr}}
}
