% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild.R
\name{rebuild}
\alias{rebuild}
\title{rebuild}
\usage{
rebuild(X)
}
\arguments{
\item{X}{A numeric matrix.}

\item{(s)pca, }{(s)pls, block.(s)pls results
that contains clustering information with molecule and cluster}
}
\value{
\describe{
  \item{One}{First item}
  \item{Two}{Second item}
}
}
\description{
A more detailed description.
}
\examples{


}
