% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{silhouette}
\alias{silhouette}
\title{silhouette}
\usage{
silhouette(dmatrix, cluster)
}
\arguments{
\item{cluster}{A numeric vector of sive ncol(X) containing the cluster
information by}

\item{X}{A numeric matrix of size NxP with feature in colnames.}
}
\description{
Compute silhouette coefficient from a matrix and clusters
}
\examples{
data <- get_demo_silhouette()
data$dist <- dmatrix.proportionality.distance(data$data)
sil <- silhouette(data$dist, data$cluster$cluster)

}
