% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCluster.R
\name{summary.getCluster}
\alias{summary.getCluster}
\title{summary.getCluster}
\usage{
\method{summary}{getCluster}(getClusterObj)
}
\arguments{
\item{getClusterObj}{}
}
\description{
summary.getCluster
}
\examples{
demo <- suppressWarnings(get_demo_cluster())
X <- demo$X[c(1,2,3), ]
res <- getUpDownCluster(X)
res_cluster <- getCluster(res)
summary.getCluster(res_cluster)
}
