% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneCluster.block.spls.R
\name{tuneCluster.block.spls}
\alias{tuneCluster.block.spls}
\title{tuneCluster.block.spls}
\usage{
tuneCluster.block.spls(X, Y = NULL, ncomp = 2,
  test.list.keepX = rep(ncol(X), ncomp), test.keepY = NULL, ...)
}
\description{
tuneCluster.block.spls
}
\examples{
demo <- suppressMessages(get_demo_cluster())
X <- list(X = demo$X, Z = demo$Z)
Y <- demo$Y
test.list.keepX <- list("X" = c(5,10,15,20), "Z" = c(2,4,6,8))
test.keepY <- c(2:5)
tune.block.spls <- tuneCluster.block.spls(X = X, Y = Y, test.list.keepX = test.list.keepX, test.keepY = test.keepY, mode = "canonical")
}
