% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneCluster.spca.R
\name{tuneCluster.spca}
\alias{tuneCluster.spca}
\title{tuneCluster.spca}
\usage{
tuneCluster.spca(X, ncomp = 2, test.keepX = rep(ncol(X), ncomp), ...)
}
\description{
tuneCluster.spca
}
\examples{
demo <- suppressMessages(get_demo_cluster())
X <- demo$X
tune.spca.res <- tuneCluster.spca(X = X, ncomp = 2, test.keepX = c(2,5,7))
plot(tune.spca.res)
plot(tune.spca.res, comp = 2)
}
