% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneCluster.spls.R
\name{tuneCluster.spls}
\alias{tuneCluster.spls}
\title{tuneCluster.spls}
\usage{
tuneCluster.spls(X, Y, ncomp = 2, test.keepX = rep(ncol(X), ncomp),
  test.keepY = rep(ncol(Y), ncomp), ...)
}
\description{
tuneCluster.spls
}
\examples{
demo <- suppressMessages(get_demo_cluster())
X <- demo$X
Y <- demo$Y
tune.spls <- tuneCluster.spls(X, Y, ncomp = 2, test.keepX = c(5,10,15,20), test.keepY <- c(2,4,6))

}
