% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCluster.R
\name{getCluster}
\alias{getCluster}
\title{getCluster}
\usage{
getCluster(x)
}
\arguments{
\item{X}{A numeric matrix.}

\item{(s)pca, }{(s)pls, block.(s)pls results
that contains clustering information with molecule and cluster}
}
\value{
\describe{
  \item{One}{First item}
  \item{Two}{Second item}
}
}
\description{
A more detailed description.
}
\examples{
demo <- suppressMessages(get_demo_cluster())
pca.cluster <- getCluster(demo$pca)
spca.cluster <- getCluster(demo$spca)
pls.cluster <- getCluster(demo$pls)
spls.cluster <- getCluster(demo$spls)
block.pls.cluster <- getCluster(demo$block.pls)
block.spls.cluster <- getCluster(demo$block.spls)

}
