% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSilhouette.R
\name{getSilhouette}
\alias{getSilhouette}
\title{getSilhouette}
\usage{
getSilhouette(object)
}
\arguments{
\item{X}{A numeric matrix.}

\item{cluster}{A data.frame
that contains clustering information with molecule and cluster}
}
\value{

}
\description{
\code{getSilhouette} is a generic function that compute silhouette coefficient
for an object of the type \code{pca}, \code{spca}, \code{pls}, \code{spls},
\code{block.pls}, \code{block.spls}.
}
\details{
This method extract the componant contribution depending on the object,
perform the clustering step, and compute the silhouette coefficient.
}
\examples{
demo <- suppressMessages(get_demo_cluster())
# pca
getSilhouette(object = demo$pca)
# spca
getSilhouette(object = demo$spca)


}
