% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUpDownCluster.R
\name{getUpDownCluster}
\alias{getUpDownCluster}
\title{Up-Down clustering}
\usage{
getUpDownCluster(X)
}
\arguments{
\item{X}{a dataframe or list of dataframe with the same number of rows.}
}
\description{
Performs a clustering based on the signs of variation between 2 timepoints.
}
\examples{
demo <- suppressWarnings(get_demo_cluster())
X <- list(X = demo$X, Y = demo$Y, Z = demo$Z)
res <- getUpDownCluster(X)
class(res)
getCluster(res)

X <- demo$X
res <- getUpDownCluster(X)
}
